error correction

+++ apt/app/src/main/java/com/example/apt/Clint/bfcp_floor_id_list.java
@@ -1,10 +1,10 @@
 package com.example.apt.Clint;
 
 public class bfcp_floor_id_list {
-        public int ID;                    // FLOOR-ID
+        public short ID;                    // FLOOR-ID
         public bfcp_floor_id_list next;   // Pointer to next FLOOR-ID instance
 
-        public bfcp_floor_id_list(int ID, bfcp_floor_id_list next) {
+        public bfcp_floor_id_list(short ID, bfcp_floor_id_list next) {
             this.ID = ID;
             this.next = next;
         }
