+++ apt/app/src/main/java/com/example/apt/Clint/bfcp_overall_request_status.java
@@ -1,11 +1,11 @@
 package com.example.apt.Clint;
 
 public class bfcp_overall_request_status {
-    public int frqID;                   // OVERALL-REQUEST-STATUS-HEADER
+    public short frqID;                   // OVERALL-REQUEST-STATUS-HEADER
     public bfcp_request_status rs;      // REQUEST-STATUS, optional
     public String sInfo;                // STATUS-INFO, optional
 
-    public bfcp_overall_request_status(int frqID, bfcp_request_status rs, String sInfo) {
+    public bfcp_overall_request_status(short frqID, bfcp_request_status rs, String sInfo) {
         this.frqID = frqID;
         this.rs = rs;
         this.sInfo = sInfo;
