+++ apt/app/src/main/java/com/example/apt/Clint/bfcp_request_status.java
@@ -1,10 +1,10 @@
 package com.example.apt.Clint;
 
 public class bfcp_request_status {
-    public int rs;  // Request Status
-    public int qp;  // Queue Position
+    public short rs;  // Request Status
+    public short qp;  // Queue Position
 
-    public bfcp_request_status(int rs, int qp) {
+    public bfcp_request_status(short rs, short qp) {
         this.rs = rs;
         this.qp = qp;
     }
